requirements.txt
Flask==2.2.1
gunicorn==20.1.0
pymysql==1.0.2
python-dotenv==0.21.0
pyjwt==2.5.0
bcrypt==4.0.0